Split bib2json to bibliography module
